#18 Replace registerClassyNavEvent in header-area.component
